requirements.txt
langchain
langchain-community
langchain-experimental
langchain-openai
langchain-qdrant
langchain-cohere
rank_bm25
qdrant-client
ipykernel
python-dotenv
langsmith[otel]
redis
testcontainers[redis]
grandalf
arize-phoenix-otel 
openinference-instrumentation-langchain
psycopg2-binary
asyncpg
prefect-aws
pandas
prefect
ragas
rapidfuzz